+++ index.html
@@ -11,5 +11,6 @@
 <main class="central">
 
 </main>
+<script src="js/main.js" type="text/javascript"></script>
 </body>
 </html>
