+++ js/getElement.js
@@ -1,6 +1,5 @@
 export const getElementFromTemplate = (htmlTemplate) => {
   const htmlContent = document.createElement(`div`);
   htmlContent.innerHTML = htmlTemplate;
-  return htmlContent.content;
-  console.log('bing!');
+  return htmlContent.firstChild;
 };
