+++ js/greeting.js
@@ -1,6 +1,7 @@
 import {getElementFromTemplate} from "./getElement";
+import {showNextPage} from "./appPageShow";
 
-const moduleGreetingElement = getElementFromTemplate(`<template id="greeting">
+const moduleGreetingElement = getElementFromTemplate(`<div id="greeting">
   <div class="greeting central--blur">
     <div class="greeting__logo"><img src="img/logo_big.png" width="201" height="89" alt="Pixel Hunter"></div>
     <h1 class="greeting__asterisk">*</h1>
@@ -24,6 +25,11 @@ const moduleGreetingElement = getElementFromTemplate(`<template id="greeting">
       <a href="https://vk.com/htmlacademy" class="social-link  social-link--vk">Вконтакте</a>
     </div>
   </footer>
-</template>`);
+</div>`);
+
+moduleGreetingElement.getElementsByClassName("greeting__continue")[0].onclick = () => {
+  showNextPage();
+};
+
 
 export {moduleGreetingElement};
